requirements.txt
requests==2.28.1
python-dotenv==0.20.0
Flask==2.2.2
Flask-RESTful==0.3.9
validators==0.22.0
nltk==3.8.1
bs4==0.0.1